requirements/dev.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alpaca-py==0.43.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
deprecation==2.1.0
executing==2.2.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gspread==6.2.1
h5py==3.15.1
httplib2==0.31.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml-dtypes==0.4.1
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.7.0
namex==0.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pefile==2023.2.7
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
sseclient-py==1.8.0
stack-data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
watchdog==6.0.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
xgboost==3.1.1
yarl==1.22.0
google-genai==1.53.0